bios support to the main branch

+++ setup.sh
@@ -1,15 +1,8 @@
 #!/bin/bash
 
 clear
-echo "UEFI Is Required, are you on UEFI? [ y / n ]" # could be removed at some point... hopefully
+echo "Are you on UEFI? [ y / n ]"
 read UEFICHECK
-if [ $UEFICHECK = n ]
-	then
-		echo "You Must use UEFI to Continue"
-        sleep 5
-		clear
-		exit
-fi
 
 clear
 
@@ -43,21 +36,42 @@ echo "Partitioning..."
 
 sleep 1
 
+if [ $UEFICHECK = n ]
+    then
+        parted /dev/$DRIVETYPE mklabel msdos
+fi
+if [ $UEFICHECK = y ]
+    then
 parted /dev/$DRIVETYPE mklabel gpt
 parted /dev/$DRIVETYPE mkpart "EFI" fat32 1MiB 301MiB
 parted /dev/$DRIVETYPE set 1 esp on
+fi
 
 clear
 echo "How much Swap Space do you want? Example: 2GiB, 512MiB"
 read SWAPAMOUNT
+if [ $UEFICHECK = y ]
+    then
+
 parted /dev/$DRIVETYPE mkpart "swap" linux-swap 301MiB $SWAPAMOUNT
 
 parted /dev/$DRIVETYPE mkpart "root" ext4 $SWAPAMOUNT 100%
 
+fi
+
+if [ $UEFICHECK = n ]
+    then
+        parted /dev/$DRIVETYPE mkpart primary linux-swap 1MiB $SWAPAMOUNT
+
+        parted /dev/$DRIVETYPE mkpart primary ext4 $SWAPAMOUNT 100%
+
+fi
 clear
 echo "Formatting Partitions..." 
 sleep 2
-if [ $DRIVETYPE = vda ]
+if [ $UEFICHECK = y ]
+    then
+    if [ $DRIVETYPE = vda ]
 	then
 		mkfs.ext4 /dev/vda3
         mkswap /dev/vda2
@@ -66,8 +80,8 @@ if [ $DRIVETYPE = vda ]
         mount /dev/vda3 /mnt 
         mount --mkdir /dev/vda1 /mnt/boot
         swapon /dev/vda2
-fi
-if [ $DRIVETYPE = sda ]
+    fi
+    if [ $DRIVETYPE = sda ]
 	then
 		mkfs.ext4 /dev/sda3
         mkswap /dev/sda2
@@ -76,8 +90,8 @@ if [ $DRIVETYPE = sda ]
         mount /dev/sda3 /mnt 
         mount --mkdir /dev/sda1 /mnt/boot
         swapon /dev/sda2
-fi
-if [ $DRIVETYPE = nvme0n1 ]
+    fi
+    if [ $DRIVETYPE = nvme0n1 ]
 	then
 		mkfs.ext4 /dev/nvme0n1p3
         mkswap /dev/nvme0n1p2
@@ -86,6 +100,43 @@ if [ $DRIVETYPE = nvme0n1 ]
         mount /dev/nvme0n1p3 /mnt 
         mount --mkdir /dev/nvme0n1p1 /mnt/boot
         swapon /dev/nvme0n1p2
+    fi
+
+fi
+
+if [ $UEFICHECK = n ]
+    then
+        if [ $DRIVETYPE = vda ]
+	then
+		mkfs.ext4 /dev/vda2
+        mkswap /dev/vda1
+        
+
+        mount /dev/vda2 /mnt 
+        
+        swapon /dev/vda1
+    fi
+    if [ $DRIVETYPE = sda ]
+	then
+		mkfs.ext4 /dev/sda2
+        mkswap /dev/sda1
+        parted /dev/sda set 2 boot on
+
+        mount /dev/sda2 /mnt 
+        
+        swapon /dev/sda1
+    fi
+    if [ $DRIVETYPE = nvme0n1 ]
+	then
+		mkfs.ext4 /dev/nvme0n1p2
+        mkswap /dev/nvme0n1p1
+        parted /dev/nvme0n1 set 2 boot on
+        
+
+        mount /dev/nvme0n1p2 /mnt 
+        
+        swapon /dev/nvme0n1p1
+    fi
 fi
 
 clear
